account-a/lambda/requirements.txt
psycopg2-binary==2.9.7
boto3==1.28.62
botocore==1.31.62
